Added FrameFactory so a Type could be assigned to a IKarassFrame. Updated KarassDependencies to return a func.
