+++ .idea/.idea.KanKan/.idea/contentModel.xml
@@ -9,6 +9,7 @@
       <e p="KanKanCore" t="IncludeRecursive">
         <e p="bin" t="ExcludeRecursive" />
         <e p="Factories" t="Include">
+          <e p="FrameFactory.cs" t="Include" />
           <e p="KarassFactory.cs" t="Include" />
         </e>
         <e p="KanKan.cs" t="Include" />
@@ -21,6 +22,7 @@
             <e p="IDependencies.cs" t="Include" />
             <e p="IKarass.cs" t="Include" />
             <e p="IKarassFactory.cs" t="Include" />
+            <e p="IKarassFrame.cs" t="Include" />
             <e p="IKarassMessage.cs" t="Include" />
           </e>
           <e p="Karass.cs" t="Include" />
@@ -50,6 +52,9 @@
           <e p="GivenMultipleFrameSetsInFrameCollection.cs" t="Include" />
           <e p="GivenOneFrameSetInFrameCollection.cs" t="Include" />
         </e>
+        <e p="FrameTests" t="Include">
+          <e p="FrameTests.cs" t="Include" />
+        </e>
         <e p="IEnumeratorTests" t="Include">
           <e p="KanKanCurrentTests.cs" t="Include" />
           <e p="KanKanResetTests.cs" t="Include" />
@@ -58,7 +63,8 @@
         <e p="KanKanTest.csproj" t="IncludeRecursive" />
         <e p="KanKanTest.csproj.DotSettings" t="Include" />
         <e p="KarassFactoryTests" t="Include">
-          <e p="KarassFactoryTests.cs" t="Include" />
+          <e p="KarassFactoryGivenAnArrayOfFrames.cs" t="Include" />
+          <e p="KarassFactoryGivenAnNestedArrayOfFrames.cs" t="Include" />
         </e>
         <e p="MessageTests" t="Include">
           <e p="KanKanMessageTests.cs" t="Include" />
@@ -67,6 +73,14 @@
         <e p="Mocks" t="Include">
           <e p="Dependencies" t="Include">
             <e p="DependenciesDummy.cs" t="Include" />
+            <e p="KarassDependenciesSpy.cs" t="Include" />
+          </e>
+          <e p="KarassFrame" t="Include">
+            <e p="FrameStruct" t="Include">
+              <e p="FrameStructDummy.cs" t="Include" />
+            </e>
+            <e p="KarassFrameDummy.cs" t="Include" />
+            <e p="KarassFrameSpy.cs" t="Include" />
           </e>
           <e p="Messages" t="Include">
             <e p="KarassMessageSpy.cs" t="Include" />
@@ -88,7 +102,7 @@
         </e>
         <e p="ProgressionTests" t="Include">
           <e p="GivenASingleFrameSet.cs" t="Include" />
-          <e p="GivenMultipleFrameSets.cs" t="Include" />
+          <e p="GivenMultipleFrameSetsWhenNextFramesIsUpdated.cs" t="Include" />
           <e p="KanKanFrameProgressionTests.cs" t="Include" />
         </e>
         <e p="SequentialKarassTests" t="Include">
@@ -105,8 +119,8 @@
       </e>
       <e p="packages" t="ExcludeRecursive" />
     </e>
-    <e p="$USER_HOME$/Library/Caches/Rider2018.2/extResources" t="IncludeRecursive" />
-    <e p="$USER_HOME$/Library/Caches/Rider2018.2/resharper-host/local/Transient/ReSharperHost/v182/SolutionCaches/_KanKan.1689740154.00" t="ExcludeRecursive" />
+    <e p="$USER_HOME$/Library/Caches/Rider2018.3/extResources" t="IncludeRecursive" />
+    <e p="$USER_HOME$/Library/Caches/Rider2018.3/resharper-host/local/Transient/ReSharperHost/v183/SolutionCaches/_KanKan.1689740154.00" t="ExcludeRecursive" />
     <e p="$APPLICATION_CONFIG_DIR$/scratches" t="IncludeRecursive" />
   </component>
 </project>
