+++ KanKanCore/Karass/Interface/IDependencies.cs
@@ -1,7 +1,11 @@
+using System;
+
 namespace KanKanCore.Karass.Interface
 {
     public interface IDependencies
     {
         T Get<T>() where T : class;
+        
+        void Register<T>(Func<dynamic> resolver);
     }
 }
