+++ KanKanTest/DependenciesTests/KarassDependenciesTests.cs
@@ -1,3 +1,4 @@
+using System;
 using KanKanCore.Karass.Dependencies;
 using KanKanCore.Karass.Interface;
 using NUnit.Framework;
@@ -14,7 +15,7 @@ namespace KanKanTest.DependenciesTests
 
         private class TestClass : ITestInterface
         {
-            private readonly string _returnValue;
+            public readonly string _returnValue;
 
             public TestClass(string returnValue)
             {
@@ -31,8 +32,10 @@ namespace KanKanTest.DependenciesTests
         public string ThenCorrectlyReturnType(string name)
         {
             KarassDependencies dependencies = new KarassDependencies();
-            dependencies.Register<ITestInterface>(new TestClass(name));
-            return dependencies.Get<ITestInterface>().Test();
+            TestClass testClass = new TestClass(name);
+            dependencies.Register<ITestInterface>(() => testClass);
+            ITestInterface t = dependencies.Get<ITestInterface>();
+            return t.Test();
         }
   
     }
