+++ KanKanTest/Mocks/Dependencies/DependenciesDummy.cs
@@ -1,3 +1,4 @@
+using System;
 using KanKanCore.Karass.Interface;
 
 namespace KanKanTest.Mocks.Dependencies
@@ -6,7 +7,12 @@ namespace KanKanTest.Mocks.Dependencies
     {
         public T Get<T>() where T : class
         {
-            throw new System.NotImplementedException();
+            throw new NotImplementedException();
+        }
+
+        public void Register<T>(Func<dynamic> resolver)
+        {
+            throw new NotImplementedException();
         }
     }
 }
