+++ KanKanTest/Mocks/KarassFrame/FrameStruct/FrameStructDummy.cs
@@ -1,0 +1,7 @@
+namespace KanKanTest.Mocks.KarassFrame.FrameStruct
+{
+    public struct FrameStructDummy
+    {
+        public string Test;
+    }
+}
